Remove unused files from lib
Remove two confusing files lib/cache.php and lib/sql.php. These files
appear to be early versions of the lib/Cache.php and lib/SQL.php files
which are actually used in the codebase. They were removed from
https://github.com/seanpringle/tendril at e838289. This removal was
apparently missed in this repo when 476741d imported other changes from
the upstream, but did so by overlaying the files from the upstream repo
rather than by applying the upstream changes as complete changes
including renames and deletions.

As a non-coincidental side effect, after this patch is merged the HEAD
of this repo should be able to be checked out on a case-insensitive file
system such as APFS.

Change-Id: Iab35db9a94f9cdafdcf619d98a682164dca2eed9