dashboard: Change tendril_purge_global_status_log_5m
tendril_purge_global_status_log_5m event is the one that is supposed to
get the table global_status_log_5m under control.
However it is impossible, it is too much data.
10 days after a purge we've saved 331934990 rows (T252331#6152001).

The event runs every 5 minutes and a quick test shows that we generate
around 220k rows in 5 minutes, so with the current DELETE limited to 10k
rows, we wouldn't even have it under control. Change that to 1M

I have done a test of deleting 1M rows and it takes around 50 seconds,
so it seems fine.

Bug: T252331
Change-Id: I3e74b60ff9fe81a17212bd6728cf3dc880140b06