dashboard.sql: Create tendril_purge_global_status_log event
Right now we are not purging the global_status_log table, which
generates around 1.5-2M rows every 5 minutes.

This event keeps the table under control, with around 370M rows.
This event has been running for around 48h now, and it has proven to be
working fine, so adding it to the repo.

Bug: T252331
Change-Id: I289766db037cd5e19abab71179392ee2a89e9208