Port use_index implementation from dbtree to tendril and apply to tree
"use_index" fix seemed to work for dbtree, port it to tendril and
use it on the /tree and /host views. This should fix both the tree
visualizations, and the list of hosts, but other views may also
require tuning.

Apparently, code for dbtree is very similar to the one found on
tendril, but it is different enough that some things have to be
changed.

Bug: T260876
Change-Id: I00a3bea46106bc58ee81dd4bc5ae83b7721210f3